seperate the buid chart in a single file
